Behavior1/requirements.txt
# Core dependencies
playwright
python-dotenv
loguru
aiohttp

# BDD testing framework
behave
allure-behave

# Database support
pymysql
psycopg2-binary

# File processing
PyYAML
openpyxl

# Test reporting
allure-python-commons

# Parallel testing (optional)
behave-parallel

# Async support
nest-asyncio

# Jenkins integration (optional)
jenkinsapi

# Document generation for UI evidence
python-docx
